refactor: group react dev overlay entry
